refactor(plugin-discord): split testing barrel into fetch.ts + index re-export
